move update logic to it's own file
